[refactor]余計なcontainerを消して統一
