--- src/containers/FlipsContainer.js
@@ -1,8 +1,0 @@
-import { connect } from 'react-redux';
-import FlipLabel from '../components/FlipLabel';
-
-const mapStateToProps = (state) => {
-  return state.user;
-};
-
-export default connect(mapStateToProps)(FlipLabel);
