--- src/containers/WorkTimeContainer.js
@@ -1,4 +1,0 @@
-import { connect } from 'react-redux';
-import WorkTimeLabel from '../components/WorkTimeLabel';
-
-export default connect()(WorkTimeLabel);
